amd-zynqmp: disable MULTILIBS when mali400 enabled
Issue: LINCD-23106

When the variable MULTILIBS is "multilib:lib32" in machine conf,
the variable ${baselib} is set to "lib64" in
oe-core/meta/conf/multilib.conf.

But the prebuilt libraries in the package libmali-xlnx is assumed
to be installed to the path "/usr/lib" of rootfs. If installed to
"/usr/lib64", they will not work.

So set MULTILIBS to be NULL when mali400 is enabled.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>