amd-zynqmp: secure-boot: add zynqmp-secureboot dtbo
Issue: LIN1025-2768

When secure-boot mode is enabled in BOOT.BIN, the registers of coresignt
are permitted to access. So we add a dtbo file zynqmp-secure-boot.dtbo
to disable coresight device node to protect from crashing of kernel.
In bsp layer, install this dtbo file to rootfs and update description in
README about how to use it.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>