amd-zynqmp: set MACHINEOVERRIDES and default XSERVER
Issue: LIN1025-3061

The XSERVER is defined in soc-zynqmp.inc of meta-xilinx layer and this
include file is only included when WRL_ZYNQMP_BOOTBIN_TARGET is enabled.
Without a correct XSERVER definition, X server will not work with sato
graphic distro. So we add XSERVER to machine conf file to ensure it can
be defined regardless WRL_ZYNQMP_BOOTBIN_TARGET is enabled.

Since "zynqmp" is added to MACHINEOVERRIDES via ${SOC_FAMILY} when
WRL_ZYNQMP_BOOTBIN_TARGET is enabled. To avoid adding it multitimes,
add "zynqmp" to MACHINEOVERRIDES by checking $SOC_FAMILY so that
xorg.conf file can be installed to rootfs.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>